Extract mock functions
