+++ controllers/apply/standard-proposal/form.test.js
@@ -1,45 +1,12 @@
 /* eslint-env jest */
 'use strict';
 const omit = require('lodash/omit');
-const random = require('lodash/random');
-const faker = require('faker');
 
+const { mockResponse } = require('./mocks');
 const formBuilder = require('./form');
 
 function mapMessages(validationResult) {
     return validationResult.messages.map(detail => detail.msg);
-}
-
-function mockResponse(overrides = {}) {
-    const defaults = {
-        projectName: 'My project',
-        projectCountries: ['england'],
-        projectLocation: 'derbyshire',
-        projectLocationDescription: 'description',
-        projectCosts: '250,000',
-        projectDurationYears: 3,
-        yourIdeaProject: faker.lorem.words(random(50, 500)),
-        yourIdeaCommunity: faker.lorem.words(random(50, 500)),
-        yourIdeaActivities: faker.lorem.words(random(50, 350)),
-        organisationLegalName: 'Example organisation',
-        organisationTradingName: 'Example trading name',
-        organisationAddress: {
-            line1: '1234 example street',
-            townCity: 'Birmingham',
-            county: 'West Midlands',
-            postcode: 'B15 1TR'
-        },
-        organisationType: 'not-for-profit-company',
-        contactName: {
-            firstName: 'Björk',
-            lastName: 'Guðmundsdóttir'
-        },
-        contactEmail: 'general.enquiries@tnlcommunityfund.org.uk',
-        contactPhone: '0345 4 10 20 30',
-        contactCommunicationNeeds: 'Large print'
-    };
-
-    return Object.assign(defaults, overrides);
 }
 
 test('empty form', () => {
@@ -164,7 +131,8 @@ test('featured messages based on allow list', () => {
         })
     });
 
-    expect(form.validation.featuredMessages.map(item => item.msg)).toEqual([
+    const messages = form.validation.featuredMessages.map(item => item.msg);
+    expect(messages).toEqual([
         expect.stringContaining('Tell us all of the locations')
     ]);
 });
